Convert GitHub Action to Shell Script
Change the inline GitHub Action shell scripting to be an explicit shell
script that lives in a new `bin/` directory.  This is done to enable
re-use of the script across CI infrastructure.  Specifically, this is done
to allow this GitHub Action to be usable in CI environments not running
GitHub Actions, e.g., Jenkins.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>